refactor: 추가 컨벤션 수정 (#ATR-95)
