--- apps/service/entities/index.ts
@@ -1,1 +1,0 @@
-export * from './article'
