+++ apps/service/features/filter-article/model/useViewType.ts
@@ -2,7 +2,7 @@
 
 import { useCallback, useState } from 'react'
 
-import { ViewType } from '@/entities'
+import { ViewType } from '@/entities/article'
 
 export default function useViewType() {
   const [viewType, setView] = useState<ViewType>('gallery')
