+++ apps/service/features/filter-article/ui/category-dropdown-btn.tsx
@@ -7,7 +7,7 @@ import {
   NewsletterCategory,
   NEWSLETTER_CATEGORY,
   useUserCategoriesQuery,
-} from '@/entities'
+} from '@/entities/article'
 import { useClickedOutsideOfElement } from '@/shared'
 
 type CategoryDropdownProps = {
