--- apps/service/features/index.ts
@@ -1,1 +1,0 @@
-export * from './filter-article'
